version_1/pyproject.toml
[tool.poetry]
name = "version-1"
version = "0.1.0"
description = ""
authors = ["seejn <seejnmaharjan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "1.14.0"
psycopg2-binary = "2.9.10"
pydantic = "2.10.3"
pydantic-settings = "2.6.1"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
